refactor(handlers): split execute.go into focused files (3169 -> 807 LOC) (#412)
Split the execute handler god file into logical units. Pure code
movement, no behavior change, no renames, no new interfaces.

Before: execute.go = 3169 lines, 74 functions
After:
  execute.go          =  807 lines (types, interfaces, controller, sync/async/status handlers)
  execute_ard.go      =  278 lines (external ARD call handling)
  execute_events.go   =  295 lines (event publishing, workflow status updates)
  execute_prepare.go  =  412 lines (execution preparation, replay, target parsing)
  execute_lifecycle.go=  460 lines (wait, callAgent, complete, fail, triggerWebhook)
  execute_helpers.go  =  999 lines (headers, resolution, webhook normalization, error classification, async pool)

All methods remain on *executionController. No import path changes
for callers. Tests pass unchanged.
