+++ control-plane/internal/handlers/execute.go
@@ -1,38 +1,22 @@
 package handlers
 
 import (
-	"bytes"
 	"context"
-	"crypto/hmac"
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"mime"
 	"net/http"
-	"net/url"
-	"os"
-	"runtime"
-	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
-	"github.com/Agent-Field/agentfield/control-plane/internal/ard"
 	"github.com/Agent-Field/agentfield/control-plane/internal/config"
 	"github.com/Agent-Field/agentfield/control-plane/internal/events"
 	"github.com/Agent-Field/agentfield/control-plane/internal/logger"
-	"github.com/Agent-Field/agentfield/control-plane/internal/server/middleware"
 	"github.com/Agent-Field/agentfield/control-plane/internal/services"
-	"github.com/Agent-Field/agentfield/control-plane/internal/utils"
 	"github.com/Agent-Field/agentfield/control-plane/pkg/types"
 
 	"github.com/gin-gonic/gin"
-	"github.com/rs/zerolog"
 )
 
 // ExecutionStore captures the storage operations required by the simplified execution handlers.
@@ -445,263 +429,6 @@ func (c *executionController) handleSync(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, response)
 }
 
-func (c *executionController) tryHandleExternalARDCall(ctx *gin.Context) bool {
-	if c.readARDConfig == nil {
-		return false
-	}
-	targetParam := strings.TrimSpace(ctx.Param("target"))
-	if !strings.HasPrefix(targetParam, "external.") {
-		return false
-	}
-
-	var req ExecuteRequest
-	if err := ctx.ShouldBindJSON(&req); err != nil {
-		RespondBadRequest(ctx, "invalid request body: "+err.Error())
-		return true
-	}
-	if req.Input == nil {
-		req.Input = map[string]interface{}{}
-	}
-
-	reader, ok := c.store.(ard.StateReader)
-	if !ok {
-		RespondInternalError(ctx, "external ARD invocation state is unavailable")
-		return true
-	}
-	state, err := ard.LoadStateReadOnly(ctx.Request.Context(), reader)
-	if err != nil {
-		RespondInternalError(ctx, err.Error())
-		return true
-	}
-	effective := ard.Effective(c.readARDConfig(), state)
-	if !effective.ExternalInvocationEnabled {
-		RespondError(ctx, http.StatusForbidden, "external ARD invocation is disabled by config")
-		return true
-	}
-
-	entry, binding, ok := externalARDBindingForTarget(state, targetParam)
-	if !ok {
-		RespondNotFound(ctx, "external ARD target is not imported and callable")
-		return true
-	}
-	if err := validateExternalARDOperation(req, binding); err != nil {
-		writeExecutionError(ctx, err)
-		return true
-	}
-
-	headers := readExecutionHeaders(ctx)
-	runID := headers.runID
-	if runID == "" {
-		runID = utils.GenerateRunID()
-	}
-	executionID := utils.GenerateExecutionID()
-	start := time.Now().UTC()
-	clientPayload := map[string]interface{}{"input": req.Input}
-	if len(req.Context) > 0 {
-		clientPayload["context"] = req.Context
-	}
-	storedPayload, err := json.Marshal(clientPayload)
-	if err != nil {
-		writeExecutionError(ctx, fmt.Errorf("encode execution payload: %w", err))
-		return true
-	}
-	externalReasonerID := strings.TrimPrefix(targetParam, "external.")
-	exec := &types.Execution{
-		ExecutionID:       executionID,
-		RunID:             runID,
-		ParentExecutionID: headers.parentExecutionID,
-		AgentNodeID:       "external",
-		ReasonerID:        externalReasonerID,
-		NodeID:            "external",
-		Status:            types.ExecutionStatusRunning,
-		InputPayload:      json.RawMessage(storedPayload),
-		InputURI:          c.savePayload(ctx.Request.Context(), storedPayload),
-		StartedAt:         start,
-		CreatedAt:         start,
-		UpdatedAt:         start,
-	}
-	if headers.sessionID != nil {
-		exec.SessionID = headers.sessionID
-	}
-	if headers.actorID != nil {
-		exec.ActorID = headers.actorID
-	}
-	if err := c.store.CreateExecutionRecord(ctx.Request.Context(), exec); err != nil {
-		writeExecutionError(ctx, fmt.Errorf("create external ARD execution record: %w", err))
-		return true
-	}
-
-	result, err := c.callExternalARD(ctx.Request.Context(), req, entry, binding, runID, executionID)
-	if err != nil {
-		_ = c.finishExternalARDExecution(ctx.Request.Context(), executionID, types.ExecutionStatusFailed, time.Since(start), nil, err)
-		writeExecutionError(ctx, err)
-		return true
-	}
-	resultBytes, err := json.Marshal(result)
-	if err != nil {
-		_ = c.finishExternalARDExecution(ctx.Request.Context(), executionID, types.ExecutionStatusFailed, time.Since(start), nil, err)
-		writeExecutionError(ctx, fmt.Errorf("encode external ARD result: %w", err))
-		return true
-	}
-	if err := c.finishExternalARDExecution(ctx.Request.Context(), executionID, types.ExecutionStatusSucceeded, time.Since(start), resultBytes, nil); err != nil {
-		writeExecutionError(ctx, err)
-		return true
-	}
-
-	ctx.Header("X-Execution-ID", executionID)
-	ctx.Header("X-Run-ID", runID)
-	ctx.JSON(http.StatusOK, ExecuteResponse{
-		ExecutionID: executionID,
-		RunID:       runID,
-		Status:      types.ExecutionStatusSucceeded,
-		Result:      result,
-		DurationMS:  time.Since(start).Milliseconds(),
-		FinishedAt:  time.Now().UTC().Format(time.RFC3339),
-	})
-	return true
-}
-
-func (c *executionController) callExternalARD(ctx context.Context, req ExecuteRequest, entry *ard.ExternalEntry, binding *ard.ExternalBinding, runID string, executionID string) (interface{}, error) {
-	endpoint := strings.TrimSpace(entry.URL)
-	if endpoint == "" {
-		return nil, &callError{statusCode: http.StatusNotImplemented, message: "external ARD entry has no URL-backed adapter"}
-	}
-	parsed, err := url.Parse(endpoint)
-	if err != nil || parsed.Host == "" || (parsed.Scheme != "http" && parsed.Scheme != "https") || parsed.User != nil {
-		return nil, &callError{statusCode: http.StatusBadGateway, message: "external ARD entry URL is invalid"}
-	}
-	if err := services.ValidateWebhookURL(endpoint); err != nil {
-		return nil, &callError{statusCode: http.StatusBadGateway, message: "external ARD entry URL rejected: " + err.Error()}
-	}
-
-	payload := map[string]interface{}{"input": req.Input}
-	if len(req.Context) > 0 {
-		payload["context"] = req.Context
-	}
-	body, err := json.Marshal(payload)
-	if err != nil {
-		return nil, fmt.Errorf("encode external ARD request: %w", err)
-	}
-
-	timeout := time.Duration(binding.TimeoutMS) * time.Millisecond
-	if timeout <= 0 {
-		timeout = 30 * time.Second
-	}
-	client := services.NewSSRFSafeClient(timeout)
-	httpReq, err := http.NewRequestWithContext(ctx, http.MethodPost, parsed.String(), bytes.NewReader(body))
-	if err != nil {
-		return nil, fmt.Errorf("create external ARD request: %w", err)
-	}
-	httpReq.Header.Set("Content-Type", "application/json")
-	httpReq.Header.Set("X-Run-ID", runID)
-	httpReq.Header.Set("X-Execution-ID", executionID)
-	httpReq.Header.Set("X-AgentField-ARD-Target", binding.LocalTarget)
-	httpReq.Header.Set("X-AgentField-ARD-Adapter", binding.Adapter)
-
-	resp, err := client.Do(httpReq)
-	if err != nil {
-		return nil, fmt.Errorf("external ARD call failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("read external ARD response: %w", err)
-	}
-	if resp.StatusCode >= http.StatusBadRequest {
-		return nil, &callError{
-			statusCode: resp.StatusCode,
-			message:    fmt.Sprintf("external ARD error (%d): %s", resp.StatusCode, truncateForLog(respBody)),
-			body:       respBody,
-		}
-	}
-	decoded := decodeJSON(respBody)
-	if envelope, ok := decoded.(map[string]interface{}); ok {
-		if result, ok := envelope["result"]; ok {
-			return result, nil
-		}
-	}
-	return decoded, nil
-}
-
-func (c *executionController) finishExternalARDExecution(ctx context.Context, executionID string, status string, elapsed time.Duration, result []byte, callErr error) error {
-	resultURI := c.savePayload(ctx, result)
-	_, err := c.store.UpdateExecutionRecord(ctx, executionID, func(current *types.Execution) (*types.Execution, error) {
-		if current == nil {
-			return nil, fmt.Errorf("execution %s not found", executionID)
-		}
-		now := time.Now().UTC()
-		current.Status = status
-		current.CompletedAt = pointerTime(now)
-		duration := elapsed.Milliseconds()
-		current.DurationMS = &duration
-		current.UpdatedAt = now
-		if len(result) > 0 {
-			current.ResultPayload = json.RawMessage(result)
-			current.ResultURI = resultURI
-		}
-		if callErr != nil {
-			errMsg := callErr.Error()
-			current.ErrorMessage = &errMsg
-			category := string(classifyExecutionError(callErr))
-			current.StatusReason = &category
-		} else {
-			current.ErrorMessage = nil
-			current.StatusReason = nil
-		}
-		return current, nil
-	})
-	if err != nil {
-		return fmt.Errorf("update external ARD execution record: %w", err)
-	}
-	return nil
-}
-
-func externalARDBindingForTarget(state ard.State, target string) (*ard.ExternalEntry, *ard.ExternalBinding, bool) {
-	for entryID, binding := range state.Bindings {
-		if !binding.Callable || strings.TrimSpace(binding.LocalTarget) != target {
-			continue
-		}
-		for i := range state.Imports {
-			if state.Imports[i].ID == entryID || state.Imports[i].ID == binding.ExternalEntryID {
-				return &state.Imports[i], &binding, true
-			}
-		}
-	}
-	return nil, nil, false
-}
-
-func validateExternalARDOperation(req ExecuteRequest, binding *ard.ExternalBinding) error {
-	if len(binding.AllowedOperations) == 0 {
-		return nil
-	}
-	contextOperation := stringValue(req.Context["operation"])
-	inputOperation := stringValue(req.Input["operation"])
-	if contextOperation != "" && inputOperation != "" && !strings.EqualFold(contextOperation, inputOperation) {
-		return &callError{statusCode: http.StatusBadRequest, message: "external ARD operation is ambiguous between input.operation and context.operation"}
-	}
-	operation := firstNonEmpty(contextOperation, inputOperation)
-	if operation == "" {
-		return &callError{statusCode: http.StatusForbidden, message: "external ARD operation is required by binding policy"}
-	}
-	for _, allowed := range binding.AllowedOperations {
-		if strings.EqualFold(strings.TrimSpace(allowed), operation) {
-			return nil
-		}
-	}
-	return &callError{statusCode: http.StatusForbidden, message: fmt.Sprintf("external ARD operation %q is not allowed by binding policy", operation)}
-}
-
-func stringValue(value interface{}) string {
-	switch typed := value.(type) {
-	case nil:
-		return ""
-	case string:
-		return strings.TrimSpace(typed)
-	default:
-		return strings.TrimSpace(fmt.Sprint(typed))
-	}
-}
 
 func (c *executionController) handleAsync(ctx *gin.Context) {
 	reqCtx := ctx.Request.Context()
@@ -1078,2092 +805,3 @@ func (c *executionController) handleStatusUpdate(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, c.renderStatusWithApproval(reqCtx, updated))
 }
 
-func (c *executionController) updateWorkflowExecutionStatus(
-	ctx context.Context,
-	executionID string,
-	status string,
-	statusReason *string,
-) {
-	if c.store == nil {
-		return
-	}
-
-	var normalizedReason *string
-	if statusReason != nil {
-		trimmed := strings.TrimSpace(*statusReason)
-		if trimmed != "" {
-			normalizedReason = &trimmed
-		}
-	}
-
-	err := c.store.UpdateWorkflowExecution(ctx, executionID, func(current *types.WorkflowExecution) (*types.WorkflowExecution, error) {
-		if current == nil {
-			return nil, fmt.Errorf("execution with ID %s not found", executionID)
-		}
-
-		current.Status = status
-		current.StatusReason = normalizedReason
-		current.UpdatedAt = time.Now().UTC()
-
-		if !types.IsTerminalExecutionStatus(status) {
-			current.CompletedAt = nil
-			if current.DurationMS != nil {
-				current.DurationMS = nil
-			}
-		}
-
-		return current, nil
-	})
-	if err != nil {
-		if strings.Contains(strings.ToLower(err.Error()), "not found") {
-			return
-		}
-		logger.Logger.Error().
-			Err(err).
-			Str("execution_id", executionID).
-			Str("status", status).
-			Msg("failed to update workflow execution status")
-	}
-}
-
-func (c *executionController) publishExecutionEvent(exec *types.Execution, status string, data map[string]interface{}) {
-	c.publishExecutionEventWithReasonerInfo(exec, status, data, nil, nil)
-}
-
-// enrichExecutionLifecycleData adds low-cardinality lifecycle dimensions used by
-// observability consumers. It does not mutate execution state or include payloads.
-func enrichExecutionLifecycleData(data map[string]interface{}, exec *types.Execution, status string) {
-	if data == nil || exec == nil {
-		return
-	}
-
-	data["is_root_execution"] = exec.ParentExecutionID == nil || strings.TrimSpace(*exec.ParentExecutionID) == ""
-	if _, ok := data["workflow_depth"]; !ok {
-		if data["is_root_execution"] == true {
-			data["workflow_depth"] = 0
-		}
-	}
-	if exec.DurationMS != nil {
-		data["duration_ms"] = *exec.DurationMS
-	}
-
-	switch status {
-	case string(types.ExecutionStatusSucceeded):
-		data["outcome"] = "succeeded"
-	case string(types.ExecutionStatusFailed):
-		data["outcome"] = "failed"
-		data["failure_category"] = canonicalFailureCategory(exec.StatusReason, "unknown")
-	case string(types.ExecutionStatusCancelled):
-		data["outcome"] = "cancelled"
-		data["failure_category"] = "cancelled"
-	case string(types.ExecutionStatusTimeout):
-		data["outcome"] = "timeout"
-		data["failure_category"] = "timeout"
-	}
-}
-
-func canonicalFailureCategory(statusReason *string, fallback string) string {
-	if statusReason == nil {
-		return fallback
-	}
-	category := strings.TrimSpace(*statusReason)
-	if separator := strings.Index(category, ":"); separator >= 0 {
-		category = strings.TrimSpace(category[:separator])
-	}
-	switch category {
-	case string(ErrorCategoryLLMUnavailable),
-		string(ErrorCategoryConcurrencyLimit),
-		string(ErrorCategoryAgentTimeout),
-		string(ErrorCategoryAgentError),
-		string(ErrorCategoryAgentUnreachable),
-		string(ErrorCategoryBadResponse),
-		string(ErrorCategoryInternal),
-		"agent_restart_orphaned",
-		"validation",
-		"permission_denied",
-		"node_unavailable",
-		"target_not_found":
-		return category
-	default:
-		return fallback
-	}
-}
-
-func (c *executionController) publishExecutionEventWithReasonerInfo(exec *types.Execution, status string, data map[string]interface{}, agent *types.AgentNode, reasonerID *string) {
-	if exec == nil {
-		return
-	}
-
-	eventType := events.ExecutionUpdated
-	switch status {
-	case string(types.ExecutionStatusSucceeded):
-		eventType = events.ExecutionCompleted
-	case string(types.ExecutionStatusFailed):
-		eventType = events.ExecutionFailed
-	case string(types.ExecutionStatusRunning):
-		eventType = events.ExecutionStarted
-	case "created":
-		eventType = events.ExecutionCreated
-	}
-
-	// Ensure data map exists
-	if data == nil {
-		data = make(map[string]interface{})
-	}
-	enrichExecutionLifecycleData(data, exec, status)
-
-	// Add reasoner_id to the event data
-	rID := exec.ReasonerID
-	if reasonerID != nil && *reasonerID != "" {
-		rID = *reasonerID
-	}
-	if rID != "" {
-		data["reasoner_id"] = rID
-	}
-
-	// Add node_id to the event data
-	if exec.NodeID != "" {
-		data["node_id"] = exec.NodeID
-	}
-	if exec.AgentNodeID != "" {
-		data["agent_node_id"] = exec.AgentNodeID
-	}
-	if exec.StatusReason != nil && *exec.StatusReason != "" {
-		data["status_reason"] = *exec.StatusReason
-		data["error_category"] = *exec.StatusReason
-	}
-	data["started_at"] = exec.StartedAt.UTC().Format(time.RFC3339)
-	if exec.CompletedAt != nil {
-		data["completed_at"] = exec.CompletedAt.UTC().Format(time.RFC3339)
-	}
-	if exec.DurationMS != nil {
-		data["duration_ms"] = *exec.DurationMS
-	}
-	if exec.SessionID != nil && *exec.SessionID != "" {
-		data["session_id"] = *exec.SessionID
-	}
-	if exec.ActorID != nil && *exec.ActorID != "" {
-		data["actor_id"] = *exec.ActorID
-	}
-	storedPayload := types.DecodeStoredExecutionPayload(exec.InputPayload)
-	if !c.redactPayloads && storedPayload.Context != nil {
-		data["context"] = storedPayload.Context
-	}
-	if workflowExec, err := c.store.GetWorkflowExecution(context.Background(), exec.ExecutionID); err == nil && workflowExec != nil {
-		data["retry_count"] = workflowExec.RetryCount
-		data["workflow_depth"] = workflowExec.WorkflowDepth
-	}
-
-	// Add reasoner definitions if agent info is available
-	if agent != nil {
-		// Find the specific reasoner being executed
-		for _, r := range agent.Reasoners {
-			if r.ID == rID {
-				data["reasoner"] = map[string]interface{}{
-					"id":            r.ID,
-					"input_schema":  r.InputSchema,
-					"output_schema": r.OutputSchema,
-				}
-				break
-			}
-		}
-
-		// Find the specific skill being executed
-		for _, s := range agent.Skills {
-			if s.ID == rID {
-				data["skill"] = map[string]interface{}{
-					"id":           s.ID,
-					"input_schema": s.InputSchema,
-					"tags":         s.Tags,
-				}
-				data["skill_id"] = s.ID
-				break
-			}
-		}
-
-		// Include all reasoners on this agent node for back-population
-		if len(agent.Reasoners) > 0 {
-			reasonerList := make([]map[string]interface{}, 0, len(agent.Reasoners))
-			for _, r := range agent.Reasoners {
-				reasonerList = append(reasonerList, map[string]interface{}{
-					"id":            r.ID,
-					"input_schema":  r.InputSchema,
-					"output_schema": r.OutputSchema,
-				})
-			}
-			data["agent_reasoners"] = reasonerList
-		}
-
-		// Include all skills on this agent node for back-population
-		if len(agent.Skills) > 0 {
-			skillList := make([]map[string]interface{}, 0, len(agent.Skills))
-			for _, s := range agent.Skills {
-				skillList = append(skillList, map[string]interface{}{
-					"id":           s.ID,
-					"input_schema": s.InputSchema,
-					"tags":         s.Tags,
-				})
-			}
-			data["agent_skills"] = skillList
-		}
-
-		// Include agent node info
-		data["agent_node"] = map[string]interface{}{
-			"id":              agent.ID,
-			"base_url":        agent.BaseURL,
-			"version":         agent.Version,
-			"deployment_type": agent.DeploymentType,
-		}
-	}
-
-	event := events.ExecutionEvent{
-		Type:        eventType,
-		ExecutionID: exec.ExecutionID,
-		WorkflowID:  exec.RunID,
-		AgentNodeID: exec.AgentNodeID,
-		Status:      status,
-		Timestamp:   time.Now(),
-		Data:        data,
-	}
-	if c.eventBus != nil {
-		c.eventBus.Publish(event)
-	}
-	events.GlobalExecutionEventBus.Publish(event)
-}
-
-// publishExecutionStartedEvent emits the ExecutionStarted event with full reasoner context
-func (c *executionController) publishExecutionStartedEvent(plan *preparedExecution) {
-	if plan == nil || plan.exec == nil {
-		return
-	}
-
-	data := map[string]interface{}{
-		"target_type":       plan.targetType,
-		"execution_mode":    plan.executionMode,
-		"transition_source": "execution_controller",
-	}
-
-	// Include input payload info (not the full payload, just metadata)
-	if len(plan.exec.InputPayload) > 0 {
-		data["input_size"] = len(plan.exec.InputPayload)
-	}
-
-	c.publishExecutionEventWithReasonerInfo(
-		plan.exec,
-		string(types.ExecutionStatusRunning),
-		data,
-		plan.agent,
-		&plan.target.TargetName,
-	)
-}
-
-// completionPollInterval is how often waitForExecutionCompletion re-reads the
-// execution record while it waits on the event bus. Var, not const, so tests
-// can shrink it.
-var completionPollInterval = 500 * time.Millisecond
-
-// waitForExecutionCompletion waits for an execution to complete by subscribing to the event bus.
-// It returns the completed execution record or an error if the execution fails or times out.
-// This is used when agents return HTTP 202 (async acknowledgment) but the sync endpoint needs to wait for completion.
-//
-// The event bus is the fast path, but not every writer of a terminal status
-// publishes a lifecycle event: the SDK's reasoner.completed workflow event
-// persists the terminal state through WorkflowExecutionEventHandler without
-// one, and when the authoritative /status callback lands afterwards it is an
-// idempotent terminal->terminal update. If that ordering wins the race, the
-// only signal a synchronous caller would ever get was the timeout, ninety
-// seconds after its result had been stored. The store is therefore polled as
-// a fallback, so the wait is bounded by completionPollInterval rather than by
-// whichever callback happened to arrive first.
-func (c *executionController) waitForExecutionCompletion(ctx context.Context, executionID string, timeout time.Duration) (*types.Execution, error) {
-	if c.eventBus == nil {
-		return nil, fmt.Errorf("event bus not available")
-	}
-
-	// Create unique subscriber ID for this wait operation
-	subscriberID := fmt.Sprintf("sync-wait-%s", executionID)
-
-	// Subscribe to events
-	eventChan := c.eventBus.Subscribe(subscriberID)
-	defer c.eventBus.Unsubscribe(subscriberID)
-
-	// Create timeout timer
-	timer := time.NewTimer(timeout)
-	defer timer.Stop()
-
-	// Fallback for terminal states that reach the store without an event.
-	poll := time.NewTicker(completionPollInterval)
-	defer poll.Stop()
-
-	logger.Logger.Debug().
-		Str("execution_id", executionID).
-		Dur("timeout", timeout).
-		Msg("waiting for execution completion via event bus")
-
-	// Check if execution already completed before we subscribed (race condition:
-	// fast agents may POST the callback before we subscribe to the event bus).
-	if existing, err := c.store.GetExecutionRecord(ctx, executionID); err == nil && existing != nil {
-		if types.IsTerminalExecutionStatus(existing.Status) {
-			logger.Logger.Debug().
-				Str("execution_id", executionID).
-				Str("status", existing.Status).
-				Msg("execution already completed before event subscription")
-			return existing, nil
-		}
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-
-		case <-timer.C:
-			logger.Logger.Warn().
-				Str("execution_id", executionID).
-				Dur("timeout", timeout).
-				Msg("execution completion timeout")
-			return nil, fmt.Errorf("execution timeout after %v", timeout)
-
-		case <-poll.C:
-			existing, err := c.store.GetExecutionRecord(ctx, executionID)
-			if err != nil || existing == nil || !types.IsTerminalExecutionStatus(existing.Status) {
-				continue
-			}
-			logger.Logger.Debug().
-				Str("execution_id", executionID).
-				Str("status", existing.Status).
-				Msg("execution reached a terminal state without a terminal event; completing from the stored record")
-			return existing, nil
-
-		case event := <-eventChan:
-			// Only process events for this specific execution
-			if event.ExecutionID != executionID {
-				continue
-			}
-
-			// Check if this is a terminal event
-			if event.Type == events.ExecutionCompleted || event.Type == events.ExecutionFailed {
-				logger.Logger.Debug().
-					Str("execution_id", executionID).
-					Str("event_type", string(event.Type)).
-					Msg("received terminal execution event")
-
-				// Fetch the updated execution record
-				exec, err := c.store.GetExecutionRecord(ctx, executionID)
-				if err != nil {
-					return nil, fmt.Errorf("failed to fetch execution after completion: %w", err)
-				}
-				if exec == nil {
-					return nil, fmt.Errorf("execution %s not found after completion event", executionID)
-				}
-
-				return exec, nil
-			}
-
-			// Continue waiting for other event types (ExecutionUpdated, etc.)
-		}
-	}
-}
-
-// waitForResume waits for a paused execution to be resumed or cancelled.
-// It returns nil when resumed and an error when cancelled or context is cancelled.
-func (c *executionController) waitForResume(ctx context.Context, executionID string) error {
-	if c.eventBus == nil {
-		return fmt.Errorf("event bus not available")
-	}
-
-	// Create unique subscriber ID for this wait operation. Include a monotonic
-	// counter so that multiple goroutines waiting on the same execution (e.g.
-	// parallel DAG branches) each get their own event channel.
-	subscriberID := fmt.Sprintf("pause-wait-%s-%d", executionID, time.Now().UnixNano())
-
-	// Subscribe to events.
-	eventChan := c.eventBus.Subscribe(subscriberID)
-	defer c.eventBus.Unsubscribe(subscriberID)
-
-	logger.Logger.Debug().
-		Str("execution_id", executionID).
-		Msg("waiting for execution resume via event bus")
-
-	// Check if execution already resumed/cancelled before we subscribed (race condition:
-	// fast status transitions may happen before we subscribe to the event bus).
-	if existing, err := c.store.GetExecutionRecord(ctx, executionID); err == nil && existing != nil {
-		if existing.Status == types.ExecutionStatusCancelled {
-			return fmt.Errorf("execution cancelled")
-		}
-		if existing.Status != types.ExecutionStatusPaused {
-			logger.Logger.Debug().
-				Str("execution_id", executionID).
-				Str("status", existing.Status).
-				Msg("execution already resumed before event subscription")
-			return nil
-		}
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-
-		case event := <-eventChan:
-			// Only process events for this specific execution.
-			if event.ExecutionID != executionID {
-				continue
-			}
-
-			switch event.Type {
-			case events.ExecutionResumed:
-				logger.Logger.Debug().
-					Str("execution_id", executionID).
-					Msg("received execution resumed event")
-				return nil
-			case events.ExecutionCancelledEvent:
-				return fmt.Errorf("execution cancelled")
-			}
-
-			// Continue waiting for other event types (ExecutionUpdated, etc.)
-		}
-	}
-}
-
-type preparedExecution struct {
-	exec              *types.Execution
-	requestBody       []byte
-	agent             *types.AgentNode
-	target            *parsedTarget
-	targetType        string
-	executionMode     string
-	llmEndpoint       string
-	webhookRegistered bool
-	webhookError      *string
-	// DID context forwarded to the target agent.
-	callerDID string
-	targetDID string
-	// Version that was selected during routing (empty if default/unversioned agent)
-	routedVersion           string
-	replaySourceRunID       string
-	replayBeforeExecutionID string
-	replayMode              string
-	replayHit               *replayHit
-}
-
-func (c *executionController) prepareExecution(ctx context.Context, ginCtx *gin.Context) (*preparedExecution, error) {
-	targetParam := ginCtx.Param("target")
-	var req ExecuteRequest
-	if err := ginCtx.ShouldBindJSON(&req); err != nil {
-		return nil, fmt.Errorf("invalid request body: %w", err)
-	}
-	return c.prepareExecutionForTarget(
-		ctx,
-		targetParam,
-		req,
-		readExecutionHeaders(ginCtx),
-		middleware.GetVerifiedCallerDID(ginCtx),
-		middleware.GetTargetDID(ginCtx),
-	)
-}
-
-func (c *executionController) prepareExecutionForTarget(ctx context.Context, targetParam string, req ExecuteRequest, headers executionHeaders, callerDID, targetDID string) (*preparedExecution, error) {
-	target, err := parseTarget(targetParam)
-	if err != nil {
-		return nil, fmt.Errorf("invalid target: %w", err)
-	}
-
-	// Allow empty input for skills/reasoners that take no parameters (issue #196).
-	if req.Input == nil {
-		req.Input = map[string]interface{}{}
-	}
-
-	var (
-		sanitizedWebhook *normalizedWebhookConfig
-		webhookError     *string
-	)
-
-	if req.Webhook != nil {
-		cfg, err := normalizeWebhookRequest(req.Webhook)
-		if err != nil {
-			errMsg := err.Error()
-			webhookError = &errMsg
-		} else if cfg != nil {
-			sanitizedWebhook = cfg
-		}
-	}
-
-	// Version-aware agent resolution:
-	// 1. Try GetAgent (default unversioned agent, version='')
-	// 2. If not found, fall back to ListAgentVersions and select via weighted round-robin
-	var agent *types.AgentNode
-	var routedVersion string
-
-	agent, err = c.store.GetAgent(ctx, target.NodeID)
-	if err != nil {
-		// GetAgent returns error for "not found" — check if versioned agents exist
-		versions, listErr := c.store.ListAgentVersions(ctx, target.NodeID)
-		if listErr != nil || len(versions) == 0 {
-			return nil, fmt.Errorf("agent '%s' not found", target.NodeID)
-		}
-		// Filter to healthy nodes
-		agent, routedVersion = selectVersionedAgent(versions)
-		if agent == nil {
-			return nil, fmt.Errorf("agent '%s' has no healthy versioned nodes", target.NodeID)
-		}
-	}
-
-	// Block calls to agents that are pending approval (e.g. tags revoked).
-	// Matches the contract used by reasoners.go / skills.go / permission
-	// middleware: stable machine code in `error`, friendly text in `message`.
-	if agent.LifecycleStatus == types.AgentStatusPendingApproval {
-		return nil, &executionPreconditionError{
-			code:      http.StatusServiceUnavailable,
-			message:   fmt.Sprintf("agent node '%s' is awaiting tag approval and cannot execute", target.NodeID),
-			category:  ErrorCategoryAgentError,
-			errorCode: "agent_pending_approval",
-		}
-	}
-
-	if agent.DeploymentType == "" && agent.Metadata.Custom != nil {
-		if v, ok := agent.Metadata.Custom["serverless"]; ok && fmt.Sprint(v) == "true" {
-			agent.DeploymentType = "serverless"
-		}
-	}
-
-	// Reject a call to a node we already know is down BEFORE the execution
-	// record is created. Dispatching into a dead node would persist a row and
-	// then fail it, charging the caller for a failed execution when nothing
-	// was ever attempted. See execute_agent_restart.go.
-	//
-	// Serverless nodes are exempt: they have no heartbeat loop and the health
-	// monitor never polls them, so the presence sweep marks every serverless
-	// node inactive shortly after registration — their recorded health says
-	// nothing about whether an invocation would succeed. Replay requests are
-	// also exempt: a replay hit is served from the recorded run without ever
-	// contacting the agent, so the node being down must not reject it (a
-	// replay miss simply dials and fails exactly as it did before this gate).
-	if agent.DeploymentType != "serverless" && strings.TrimSpace(headers.replaySourceRunID) == "" {
-		if err := ensureAgentDispatchable(agent, target.NodeID); err != nil {
-			return nil, err
-		}
-	}
-	if agent.DeploymentType == "serverless" && (agent.InvocationURL == nil || strings.TrimSpace(*agent.InvocationURL) == "") {
-		if trimmed := strings.TrimSpace(agent.BaseURL); trimmed != "" {
-			execURL := strings.TrimSuffix(trimmed, "/") + "/execute"
-			agent.InvocationURL = &execURL
-		}
-	}
-
-	targetType, err := determineTargetType(agent, target.TargetName)
-	if err != nil {
-		return nil, err
-	}
-	target.TargetType = targetType
-
-	runID := headers.runID
-	if runID == "" {
-		runID = utils.GenerateRunID()
-	}
-
-	executionID := utils.GenerateExecutionID()
-	now := time.Now().UTC()
-
-	clientPayload := map[string]interface{}{
-		"input": req.Input,
-	}
-	if len(req.Context) > 0 {
-		clientPayload["context"] = req.Context
-	}
-
-	storedPayload, err := json.Marshal(clientPayload)
-	if err != nil {
-		return nil, fmt.Errorf("encode execution payload: %w", err)
-	}
-
-	exec := &types.Execution{
-		ExecutionID:       executionID,
-		RunID:             runID,
-		ParentExecutionID: headers.parentExecutionID,
-		AgentNodeID:       agent.ID,
-		ReasonerID:        target.TargetName,
-		NodeID:            target.NodeID,
-		Status:            types.ExecutionStatusRunning,
-		InputPayload:      json.RawMessage(storedPayload),
-		StartedAt:         now,
-		CreatedAt:         now,
-		UpdatedAt:         now,
-	}
-
-	agentPayload := make(map[string]interface{}, len(req.Input))
-	for key, value := range req.Input {
-		agentPayload[key] = value
-	}
-
-	var agentPayloadBytes []byte
-	if agent.DeploymentType == "serverless" {
-		agentPayloadBytes, err = json.Marshal(buildServerlessPayload(target, exec, headers, agentPayload))
-	} else {
-		agentPayloadBytes, err = json.Marshal(agentPayload)
-	}
-	if err != nil {
-		return nil, fmt.Errorf("encode agent payload: %w", err)
-	}
-
-	inputURI := c.savePayload(ctx, storedPayload)
-	exec.InputURI = inputURI
-
-	if headers.sessionID != nil {
-		exec.SessionID = headers.sessionID
-	}
-	if headers.actorID != nil {
-		exec.ActorID = headers.actorID
-	}
-
-	if err := c.store.CreateExecutionRecord(ctx, exec); err != nil {
-		return nil, fmt.Errorf("create execution record: %w", err)
-	}
-
-	var webhookRegistered bool
-	if sanitizedWebhook != nil && webhookError == nil {
-		registration := &types.ExecutionWebhook{
-			ExecutionID:   executionID,
-			URL:           sanitizedWebhook.URL,
-			Headers:       sanitizedWebhook.Headers,
-			Status:        types.ExecutionWebhookStatusPending,
-			AttemptCount:  0,
-			NextAttemptAt: pointerTime(now),
-		}
-		if sanitizedWebhook.Secret != nil {
-			registration.Secret = sanitizedWebhook.Secret
-		}
-		if err := c.store.RegisterExecutionWebhook(ctx, registration); err != nil {
-			logger.Logger.Error().Err(err).Str("execution_id", executionID).Msg("failed to register execution webhook")
-			errMsg := err.Error()
-			webhookError = &errMsg
-		} else {
-			webhookRegistered = true
-			exec.WebhookRegistered = true
-		}
-	}
-
-	if !webhookRegistered {
-		exec.WebhookRegistered = false
-	}
-
-	c.ensureWorkflowExecutionRecord(ctx, exec, target, storedPayload)
-
-	hit, err := c.findReplayHit(ctx, headers, target, storedPayload)
-	if err != nil {
-		return nil, err
-	}
-
-	return &preparedExecution{
-		exec:                    exec,
-		requestBody:             agentPayloadBytes,
-		agent:                   agent,
-		target:                  target,
-		targetType:              targetType,
-		llmEndpoint:             extractRequestedLLMEndpoint(req),
-		webhookRegistered:       webhookRegistered,
-		webhookError:            webhookError,
-		callerDID:               callerDID,
-		targetDID:               targetDID,
-		routedVersion:           routedVersion,
-		replaySourceRunID:       headers.replaySourceRunID,
-		replayBeforeExecutionID: headers.replayBeforeExecutionID,
-		replayMode:              headers.replayMode,
-		replayHit:               hit,
-	}, nil
-}
-
-// findReplayHit returns a previously-succeeded child output to reuse for the
-// current app.call, or nil to run it normally. Only child executions (those with
-// a parent) are eligible — the restarted root always re-runs.
-//
-// Matching is keyed solely on (node id, reasoner id, canonical input+context);
-// among matches the earliest-started succeeded source execution wins. This is
-// intentionally position- and ordering-agnostic, so two calls to the same
-// reasoner with identical input+context within a run will both reuse the first
-// source result. That is correct for deterministic graphs; callers that need a
-// distinct result per identical call should vary the input/context or restart
-// with reuse=none.
-func (c *executionController) findReplayHit(ctx context.Context, headers executionHeaders, target *parsedTarget, storedPayload []byte) (*replayHit, error) {
-	if target == nil || headers.parentExecutionID == nil {
-		return nil, nil
-	}
-	sourceRunID := strings.TrimSpace(headers.replaySourceRunID)
-	if sourceRunID == "" {
-		return nil, nil
-	}
-	mode := strings.TrimSpace(headers.replayMode)
-	if mode == "" {
-		mode = "succeeded-before"
-	}
-	if mode == "none" {
-		return nil, nil
-	}
-	if mode != "succeeded-before" && mode != "all-succeeded" {
-		return nil, fmt.Errorf("unsupported replay mode %q", mode)
-	}
-
-	executions, err := c.store.QueryExecutionRecords(ctx, types.ExecutionFilter{
-		RunID:          &sourceRunID,
-		SortBy:         "started_at",
-		SortDescending: false,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("query replay source run: %w", err)
-	}
-	if len(executions) == 0 {
-		return nil, nil
-	}
-
-	var beforeTime *time.Time
-	if mode == "succeeded-before" && strings.TrimSpace(headers.replayBeforeExecutionID) != "" {
-		for _, exec := range executions {
-			if exec != nil && exec.ExecutionID == headers.replayBeforeExecutionID {
-				t := exec.StartedAt
-				beforeTime = &t
-				break
-			}
-		}
-		if beforeTime == nil {
-			return nil, nil
-		}
-	}
-
-	newKey, ok := canonicalReplayPayload(storedPayload)
-	if !ok {
-		return nil, nil
-	}
-	for _, exec := range executions {
-		if exec == nil {
-			continue
-		}
-		if beforeTime != nil && !exec.StartedAt.Before(*beforeTime) {
-			continue
-		}
-		if exec.Status != types.ExecutionStatusSucceeded {
-			continue
-		}
-		if exec.NodeID != target.NodeID || exec.ReasonerID != target.TargetName {
-			continue
-		}
-		if len(exec.ResultPayload) == 0 {
-			continue
-		}
-		oldKey, oldOK := canonicalReplayPayload(exec.InputPayload)
-		if !oldOK || oldKey != newKey {
-			continue
-		}
-		return &replayHit{
-			SourceExecutionID: exec.ExecutionID,
-			SourceRunID:       exec.RunID,
-			Result:            json.RawMessage(cloneBytes(exec.ResultPayload)),
-		}, nil
-	}
-	return nil, nil
-}
-
-func canonicalReplayPayload(raw []byte) (string, bool) {
-	if len(raw) == 0 {
-		return "", false
-	}
-	var v interface{}
-	if err := json.Unmarshal(raw, &v); err != nil {
-		return "", false
-	}
-	encoded, err := json.Marshal(v)
-	if err != nil {
-		return "", false
-	}
-	return string(encoded), true
-}
-
-func (c *executionController) completeReplayHit(ctx context.Context, plan *preparedExecution) error {
-	if plan == nil || plan.exec == nil || plan.replayHit == nil {
-		return fmt.Errorf("missing replay execution plan")
-	}
-	reason := "replayed_from_execution:" + plan.replayHit.SourceExecutionID
-	now := time.Now().UTC()
-	duration := int64(0)
-	result := cloneBytes(plan.replayHit.Result)
-	resultURI := c.savePayload(ctx, result)
-
-	updated, err := c.store.UpdateExecutionRecord(ctx, plan.exec.ExecutionID, func(current *types.Execution) (*types.Execution, error) {
-		if current == nil {
-			return nil, fmt.Errorf("execution %s not found", plan.exec.ExecutionID)
-		}
-		current.Status = types.ExecutionStatusSucceeded
-		current.StatusReason = &reason
-		current.ResultPayload = json.RawMessage(result)
-		current.ResultURI = resultURI
-		current.ErrorMessage = nil
-		current.CompletedAt = &now
-		current.DurationMS = &duration
-		current.UpdatedAt = now
-		return current, nil
-	})
-	if err != nil {
-		return err
-	}
-
-	c.updateWorkflowExecutionFinalState(ctx, plan.exec.ExecutionID, types.ExecutionStatusSucceeded, result, 0, nil)
-	c.updateWorkflowExecutionStatus(ctx, plan.exec.ExecutionID, types.ExecutionStatusSucceeded, &reason)
-	if plan.webhookRegistered || (updated != nil && updated.WebhookRegistered) {
-		c.triggerWebhook(plan.exec.ExecutionID)
-	}
-
-	eventData := map[string]interface{}{
-		"target_type":       plan.targetType,
-		"execution_mode":    plan.executionMode,
-		"transition_source": "replay",
-		"replay": map[string]interface{}{
-			"source_execution_id": plan.replayHit.SourceExecutionID,
-			"source_run_id":       plan.replayHit.SourceRunID,
-		},
-	}
-	if !c.redactPayloads {
-		eventData["result"] = decodeJSON(result)
-		if inputPayload := decodeJSON(plan.exec.InputPayload); inputPayload != nil {
-			eventData["input"] = inputPayload
-		}
-	}
-	c.publishExecutionEventWithReasonerInfo(updated, string(types.ExecutionStatusSucceeded), eventData, plan.agent, &plan.target.TargetName)
-	return nil
-}
-
-func extractRequestedLLMEndpoint(req ExecuteRequest) string {
-	for _, key := range []string{"llm_endpoint", "llm_backend", "backend", "provider", "model_provider"} {
-		if value, ok := req.Context[key]; ok {
-			if endpoint := strings.TrimSpace(fmt.Sprint(value)); endpoint != "" {
-				return endpoint
-			}
-		}
-	}
-	return ""
-}
-
-func (c *executionController) callAgent(ctx context.Context, plan *preparedExecution) ([]byte, time.Duration, bool, error) {
-	start := time.Now()
-
-	if plan.target != nil && plan.exec != nil {
-		PublishExecutionLog(plan.exec.ExecutionID, plan.exec.RunID, plan.target.NodeID,
-			"info", "calling agent", map[string]interface{}{
-				"agent":    plan.target.NodeID,
-				"reasoner": plan.target.TargetName,
-				"base_url": plan.agent.BaseURL,
-			})
-	}
-
-	// Check execution state before calling agent.
-	currentExec, err := c.store.GetExecutionRecord(ctx, plan.exec.ExecutionID)
-	if err == nil && currentExec != nil {
-		if currentExec.Status == types.ExecutionStatusCancelled {
-			return nil, 0, false, fmt.Errorf("execution cancelled")
-		}
-		if currentExec.Status == types.ExecutionStatusPaused {
-			if err := c.waitForResume(ctx, plan.exec.ExecutionID); err != nil {
-				return nil, 0, false, fmt.Errorf("execution paused and then cancelled or timed out: %w", err)
-			}
-		}
-	}
-
-	resp, err := c.dispatchAgentRequest(ctx, plan)
-	if err != nil {
-		return nil, time.Since(start), false, fmt.Errorf("agent call failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	url := buildAgentURL(plan.agent, plan.target)
-	if resp.StatusCode == http.StatusAccepted {
-		logger.Logger.Info().
-			Str("execution_id", plan.exec.ExecutionID).
-			Str("agent", plan.target.NodeID).
-			Str("reasoner", plan.target.TargetName).
-			Msg("agent acknowledged async execution")
-		return nil, time.Since(start), true, nil
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, time.Since(start), false, fmt.Errorf("read agent response: %w", err)
-	}
-
-	if plan.agent.DeploymentType == "serverless" {
-		annotateBodyForLog(
-			logger.Logger.Debug().
-				Str("agent", plan.target.NodeID).
-				Str("reasoner", plan.target.TargetName).
-				Str("url", url).
-				Int("status", resp.StatusCode),
-			resp.Header.Get("Content-Type"),
-			body,
-			c.redactPayloads,
-		).Msg("serverless response")
-	}
-
-	if resp.StatusCode >= http.StatusBadRequest {
-		return body, time.Since(start), false, &callError{
-			statusCode: resp.StatusCode,
-			message:    fmt.Sprintf("agent error (%d): %s", resp.StatusCode, truncateForLog(body)),
-			body:       body,
-		}
-	}
-
-	return body, time.Since(start), false, nil
-}
-
-func (c *executionController) completeExecution(ctx context.Context, plan *preparedExecution, result []byte, elapsed time.Duration) error {
-	if plan.target != nil && plan.exec != nil {
-		PublishExecutionLog(plan.exec.ExecutionID, plan.exec.RunID, plan.target.NodeID,
-			"info", "execution completed", map[string]interface{}{
-				"duration_ms": elapsed.Milliseconds(),
-			})
-	}
-
-	resultURI := c.savePayload(ctx, result)
-
-	var lastErr error
-	var alreadyCancelled bool
-	for attempt := 0; attempt < 5; attempt++ {
-		updated, err := c.store.UpdateExecutionRecord(ctx, plan.exec.ExecutionID, func(current *types.Execution) (*types.Execution, error) {
-			if current == nil {
-				return nil, fmt.Errorf("execution %s not found", plan.exec.ExecutionID)
-			}
-			// Guard: don't overwrite if already cancelled (e.g. by approval rejection webhook)
-			// or waiting for approval — the approval webhook handler manages the transition.
-			if current.Status == types.ExecutionStatusCancelled || current.Status == types.ExecutionStatusWaiting {
-				logger.Logger.Info().
-					Str("execution_id", plan.exec.ExecutionID).
-					Str("current_status", string(current.Status)).
-					Msg("skipping completion update; execution already cancelled or waiting for approval")
-				alreadyCancelled = true
-				return current, nil
-			}
-			now := time.Now().UTC()
-			current.Status = types.ExecutionStatusSucceeded
-			current.ResultPayload = json.RawMessage(result)
-			current.ErrorMessage = nil
-			current.CompletedAt = pointerTime(now)
-			duration := elapsed.Milliseconds()
-			current.DurationMS = &duration
-			current.UpdatedAt = now
-			current.ResultURI = resultURI
-			return current, nil
-		})
-		if err == nil {
-			if alreadyCancelled {
-				return nil
-			}
-			c.updateWorkflowExecutionFinalState(
-				ctx,
-				plan.exec.ExecutionID,
-				types.ExecutionStatusSucceeded,
-				result,
-				elapsed,
-				nil,
-			)
-			if plan.webhookRegistered || (updated != nil && updated.WebhookRegistered) {
-				c.triggerWebhook(plan.exec.ExecutionID)
-			}
-			eventData := map[string]interface{}{
-				"target_type":       plan.targetType,
-				"execution_mode":    plan.executionMode,
-				"transition_source": "execution_controller",
-			}
-			if !c.redactPayloads {
-				if payload := decodeJSON(result); payload != nil {
-					eventData["result"] = payload
-				}
-				if inputPayload := decodeJSON(plan.exec.InputPayload); inputPayload != nil {
-					eventData["input"] = inputPayload
-				}
-			}
-			c.publishExecutionEventWithReasonerInfo(updated, string(types.ExecutionStatusSucceeded), eventData, plan.agent, &plan.target.TargetName)
-			return nil
-		}
-		lastErr = err
-		if isRetryableDBError(err) {
-			time.Sleep(backoffDelay(attempt))
-			continue
-		}
-		return err
-	}
-	return lastErr
-}
-
-func (c *executionController) failExecution(ctx context.Context, plan *preparedExecution, callErr error, elapsed time.Duration, result []byte) error {
-	// Classify the error for user-facing diagnostics
-	category := classifyExecutionError(callErr)
-
-	if plan.target != nil && plan.exec != nil {
-		PublishExecutionLog(plan.exec.ExecutionID, plan.exec.RunID, plan.target.NodeID,
-			"error", "execution failed", map[string]interface{}{
-				"error":          callErr.Error(),
-				"error_category": string(category),
-				"duration_ms":    elapsed.Milliseconds(),
-			})
-	}
-
-	errMsg := callErr.Error()
-	resultURI := c.savePayload(ctx, result)
-	var lastErr error
-	var alreadyCancelled bool
-	for attempt := 0; attempt < 5; attempt++ {
-		updated, err := c.store.UpdateExecutionRecord(ctx, plan.exec.ExecutionID, func(current *types.Execution) (*types.Execution, error) {
-			if current == nil {
-				return nil, fmt.Errorf("execution %s not found", plan.exec.ExecutionID)
-			}
-			// Guard: don't overwrite if already cancelled (e.g. by approval rejection webhook)
-			// or waiting for approval — the approval webhook handler manages the transition.
-			if current.Status == types.ExecutionStatusCancelled || current.Status == types.ExecutionStatusWaiting {
-				logger.Logger.Info().
-					Str("execution_id", plan.exec.ExecutionID).
-					Str("current_status", string(current.Status)).
-					Msg("skipping failure update; execution already cancelled or waiting for approval")
-				alreadyCancelled = true
-				return current, nil
-			}
-			now := time.Now().UTC()
-			current.Status = types.ExecutionStatusFailed
-			current.ErrorMessage = &errMsg
-			categoryStr := string(category)
-			current.StatusReason = &categoryStr
-			current.CompletedAt = pointerTime(now)
-			duration := elapsed.Milliseconds()
-			current.DurationMS = &duration
-			current.UpdatedAt = now
-			if len(result) > 0 {
-				current.ResultPayload = json.RawMessage(result)
-			}
-			current.ResultURI = resultURI
-			return current, nil
-		})
-		if err == nil {
-			if alreadyCancelled {
-				return nil
-			}
-			c.updateWorkflowExecutionFinalState(
-				ctx,
-				plan.exec.ExecutionID,
-				types.ExecutionStatusFailed,
-				result,
-				elapsed,
-				&errMsg,
-			)
-			if plan.webhookRegistered || (updated != nil && updated.WebhookRegistered) {
-				c.triggerWebhook(plan.exec.ExecutionID)
-			}
-			eventData := map[string]interface{}{
-				"error":             errMsg,
-				"target_type":       plan.targetType,
-				"execution_mode":    plan.executionMode,
-				"failure_category":  string(category),
-				"transition_source": "execution_controller",
-			}
-			if !c.redactPayloads {
-				if payload := decodeJSON(result); payload != nil {
-					eventData["result"] = payload
-				}
-				if inputPayload := decodeJSON(plan.exec.InputPayload); inputPayload != nil {
-					eventData["input"] = inputPayload
-				}
-			}
-			c.publishExecutionEventWithReasonerInfo(updated, string(types.ExecutionStatusFailed), eventData, plan.agent, &plan.target.TargetName)
-			return nil
-		}
-		lastErr = err
-		if isRetryableDBError(err) {
-			time.Sleep(backoffDelay(attempt))
-			continue
-		}
-		return err
-	}
-	return lastErr
-}
-
-func (c *executionController) triggerWebhook(executionID string) {
-	if c.webhooks == nil || executionID == "" {
-		return
-	}
-	if err := c.webhooks.Notify(context.Background(), executionID); err != nil {
-		logger.Logger.Warn().Err(err).Str("execution_id", executionID).Msg("failed to enqueue webhook delivery")
-	}
-}
-
-type executionHeaders struct {
-	runID                   string
-	parentExecutionID       *string
-	sessionID               *string
-	actorID                 *string
-	replaySourceRunID       string
-	replayBeforeExecutionID string
-	replayMode              string
-}
-
-func readExecutionHeaders(ctx *gin.Context) executionHeaders {
-	runID := strings.TrimSpace(ctx.GetHeader("X-Run-ID"))
-	parent := strings.TrimSpace(ctx.GetHeader("X-Parent-Execution-ID"))
-	session := strings.TrimSpace(ctx.GetHeader("X-Session-ID"))
-	actor := strings.TrimSpace(ctx.GetHeader("X-Actor-ID"))
-	replaySourceRunID := strings.TrimSpace(ctx.GetHeader("X-AgentField-Replay-Source-Run-ID"))
-	replayBeforeExecutionID := strings.TrimSpace(ctx.GetHeader("X-AgentField-Replay-Before-Execution-ID"))
-	replayMode := strings.TrimSpace(ctx.GetHeader("X-AgentField-Replay-Mode"))
-
-	var parentPtr *string
-	if parent != "" {
-		parentPtr = &parent
-	}
-
-	var sessionPtr *string
-	if session != "" {
-		sessionPtr = &session
-	}
-
-	var actorPtr *string
-	if actor != "" {
-		actorPtr = &actor
-	}
-
-	return executionHeaders{
-		runID:                   runID,
-		parentExecutionID:       parentPtr,
-		sessionID:               sessionPtr,
-		actorID:                 actorPtr,
-		replaySourceRunID:       replaySourceRunID,
-		replayBeforeExecutionID: replayBeforeExecutionID,
-		replayMode:              replayMode,
-	}
-}
-
-type parsedTarget struct {
-	NodeID     string
-	TargetName string
-	TargetType string
-}
-
-func parseTarget(value string) (*parsedTarget, error) {
-	if value == "" {
-		return nil, errors.New("target is required")
-	}
-	parts := strings.Split(value, ".")
-	if len(parts) != 2 {
-		return nil, fmt.Errorf("target must be in format 'node_id.reasoner_name'")
-	}
-	return &parsedTarget{
-		NodeID:     parts[0],
-		TargetName: parts[1],
-	}, nil
-}
-
-func determineTargetType(agent *types.AgentNode, name string) (string, error) {
-	for _, reasoner := range agent.Reasoners {
-		if reasoner.ID == name {
-			return "reasoner", nil
-		}
-	}
-	for _, skill := range agent.Skills {
-		if skill.ID == name {
-			return "skill", nil
-		}
-	}
-	return "", fmt.Errorf("target '%s' not found on agent '%s'", name, agent.ID)
-}
-
-func buildAgentURL(agent *types.AgentNode, target *parsedTarget) string {
-	if agent == nil {
-		return ""
-	}
-	if agent.InvocationURL != nil && *agent.InvocationURL != "" {
-		return *agent.InvocationURL
-	}
-	if agent.DeploymentType == "serverless" {
-		base := strings.TrimSuffix(agent.BaseURL, "/")
-		if base == "" {
-			return ""
-		}
-		return fmt.Sprintf("%s/execute", base)
-	}
-
-	base := strings.TrimSuffix(agent.BaseURL, "/")
-	if target.TargetType == "skill" {
-		return fmt.Sprintf("%s/skills/%s", base, target.TargetName)
-	}
-	return fmt.Sprintf("%s/reasoners/%s", base, target.TargetName)
-}
-
-// versionRoundRobinCounter is used for round-robin selection across versioned agents.
-var versionRoundRobinCounter uint64
-
-// selectVersionedAgent picks a healthy agent from the versioned list using
-// weighted round-robin. Returns the selected agent and its version string.
-func selectVersionedAgent(versions []*types.AgentNode) (*types.AgentNode, string) {
-	// Filter to healthy nodes
-	var healthy []*types.AgentNode
-	for _, v := range versions {
-		if v.HealthStatus == types.HealthStatusActive && v.LifecycleStatus == types.AgentStatusReady {
-			healthy = append(healthy, v)
-		}
-	}
-	if len(healthy) == 0 {
-		// Fallback: accept any non-offline, non-pending-approval node
-		for _, v := range versions {
-			if v.LifecycleStatus != types.AgentStatusOffline && v.LifecycleStatus != types.AgentStatusPendingApproval {
-				healthy = append(healthy, v)
-			}
-		}
-	}
-	if len(healthy) == 0 {
-		return nil, ""
-	}
-
-	// Check if all weights are equal (use simple round-robin)
-	allEqual := true
-	firstWeight := healthy[0].TrafficWeight
-	totalWeight := 0
-	for _, v := range healthy {
-		w := v.TrafficWeight
-		if w <= 0 {
-			w = 100
-		}
-		totalWeight += w
-		if w != firstWeight {
-			allEqual = false
-		}
-	}
-
-	if allEqual || totalWeight == 0 {
-		// Simple round-robin
-		n := atomic.AddUint64(&versionRoundRobinCounter, 1) - 1
-		idx := n % uint64(len(healthy))
-		selected := healthy[idx]
-		return selected, selected.Version
-	}
-
-	// Weighted selection
-	n := atomic.AddUint64(&versionRoundRobinCounter, 1) - 1
-	counter := n % uint64(totalWeight)
-	cumulative := 0
-	for _, v := range healthy {
-		w := v.TrafficWeight
-		if w <= 0 {
-			w = 100
-		}
-		cumulative += w
-		if uint64(cumulative) > counter {
-			return v, v.Version
-		}
-	}
-
-	// Fallback
-	return healthy[0], healthy[0].Version
-}
-
-func buildServerlessPayload(target *parsedTarget, exec *types.Execution, headers executionHeaders, input map[string]interface{}) map[string]interface{} {
-	if target == nil || exec == nil {
-		return map[string]interface{}{
-			"input": input,
-		}
-	}
-
-	execCtx := map[string]interface{}{
-		"execution_id": exec.ExecutionID,
-		"run_id":       exec.RunID,
-		"workflow_id":  exec.RunID,
-	}
-
-	if headers.parentExecutionID != nil && *headers.parentExecutionID != "" {
-		execCtx["parent_execution_id"] = *headers.parentExecutionID
-	}
-	if headers.sessionID != nil && *headers.sessionID != "" {
-		execCtx["session_id"] = *headers.sessionID
-	}
-	if headers.actorID != nil && *headers.actorID != "" {
-		execCtx["actor_id"] = *headers.actorID
-	}
-
-	payload := map[string]interface{}{
-		"path":              fmt.Sprintf("/execute/%s", target.TargetName),
-		"target":            target.TargetName,
-		"reasoner":          target.TargetName,
-		"input":             input,
-		"execution_context": execCtx,
-	}
-
-	if target.TargetType != "" {
-		payload["type"] = target.TargetType
-		if target.TargetType == "skill" {
-			payload["skill"] = target.TargetName
-		}
-	}
-
-	return payload
-}
-
-type normalizedWebhookConfig struct {
-	URL     string
-	Secret  *string
-	Headers map[string]string
-}
-
-func normalizeWebhookRequest(req *WebhookRequest) (*normalizedWebhookConfig, error) {
-	if req == nil {
-		return nil, nil
-	}
-
-	trimmedURL := strings.TrimSpace(req.URL)
-	if trimmedURL == "" {
-		return nil, fmt.Errorf("webhook.url is required")
-	}
-
-	parsed, err := url.Parse(trimmedURL)
-	if err != nil {
-		return nil, fmt.Errorf("invalid webhook url: %w", err)
-	}
-	if parsed.Scheme == "" || parsed.Host == "" {
-		return nil, fmt.Errorf("webhook url must include scheme and host")
-	}
-	switch strings.ToLower(parsed.Scheme) {
-	case "https", "http":
-	default:
-		return nil, fmt.Errorf("webhook url must use http or https")
-	}
-	if parsed.User != nil {
-		return nil, fmt.Errorf("webhook url must not contain embedded credentials")
-	}
-	if err := services.ValidateWebhookURL(trimmedURL); err != nil {
-		return nil, fmt.Errorf("webhook url rejected: %w", err)
-	}
-	parsed.Fragment = ""
-
-	normalizedHeaders := make(map[string]string)
-	if len(req.Headers) > 0 {
-		for key, value := range req.Headers {
-			trimmedKey := strings.TrimSpace(key)
-			trimmedValue := strings.TrimSpace(value)
-			if trimmedKey == "" {
-				continue
-			}
-			if len(normalizedHeaders) >= maxWebhookHeaders {
-				return nil, fmt.Errorf("webhook.headers supports at most %d entries", maxWebhookHeaders)
-			}
-			if len(trimmedKey) > maxWebhookHeaderLength {
-				return nil, fmt.Errorf("webhook header name '%s' is too long", trimmedKey)
-			}
-			if len(trimmedValue) > maxWebhookHeaderLength {
-				return nil, fmt.Errorf("webhook header '%s' value is too long", trimmedKey)
-			}
-			normalizedHeaders[trimmedKey] = trimmedValue
-		}
-	}
-
-	var secretPtr *string
-	if trimmedSecret := strings.TrimSpace(req.Secret); trimmedSecret != "" {
-		if len(trimmedSecret) > maxWebhookSecretLength {
-			return nil, fmt.Errorf("webhook secret exceeds %d characters", maxWebhookSecretLength)
-		}
-		secretCopy := trimmedSecret
-		secretPtr = &secretCopy
-	}
-
-	return &normalizedWebhookConfig{
-		URL:     parsed.String(),
-		Secret:  secretPtr,
-		Headers: normalizedHeaders,
-	}, nil
-}
-
-func decodeJSON(payload []byte) interface{} {
-	if len(payload) == 0 {
-		return nil
-	}
-	var v interface{}
-	if err := json.Unmarshal(payload, &v); err == nil {
-		return v
-	}
-	return string(payload)
-}
-
-func renderStatus(exec *types.Execution) ExecutionStatusResponse {
-	var completedAt *string
-	if exec.CompletedAt != nil {
-		formatted := exec.CompletedAt.UTC().Format(time.RFC3339)
-		completedAt = &formatted
-	}
-
-	resp := ExecutionStatusResponse{
-		ExecutionID:       exec.ExecutionID,
-		RunID:             exec.RunID,
-		Status:            exec.Status,
-		StatusReason:      exec.StatusReason,
-		Result:            decodeJSON(exec.ResultPayload),
-		Error:             exec.ErrorMessage,
-		StartedAt:         exec.StartedAt.UTC().Format(time.RFC3339),
-		CompletedAt:       completedAt,
-		DurationMS:        exec.DurationMS,
-		WebhookRegistered: exec.WebhookRegistered,
-		WebhookEvents:     exec.WebhookEvents,
-	}
-	// For failed executions, expose the agent's raw response as error_details
-	// so callers can access structured error data (e.g., permission_denied fields).
-	if exec.Status == types.ExecutionStatusFailed && len(exec.ResultPayload) > 0 {
-		resp.ErrorDetails = decodeJSON(exec.ResultPayload)
-	}
-	return resp
-}
-
-// renderStatusWithApproval enriches the base status response with approval
-// fields from the corresponding WorkflowExecution record, if one exists.
-func (c *executionController) renderStatusWithApproval(ctx context.Context, exec *types.Execution) ExecutionStatusResponse {
-	resp := renderStatus(exec)
-
-	// Resolve webhook_registered from the execution_webhooks table since the
-	// field is not persisted on the execution record itself (db:"-").
-	if hasWH, err := c.store.HasExecutionWebhook(ctx, exec.ExecutionID); err == nil {
-		resp.WebhookRegistered = hasWH
-	}
-
-	// Best-effort enrichment — if the lookup fails we still return the base response.
-	wfExec, err := c.store.GetWorkflowExecution(ctx, exec.ExecutionID)
-	if err != nil || wfExec == nil {
-		return resp
-	}
-
-	resp.ApprovalRequestID = wfExec.ApprovalRequestID
-	resp.ApprovalStatus = wfExec.ApprovalStatus
-	resp.ApprovalRequestURL = wfExec.ApprovalRequestURL
-	return resp
-}
-
-func (c *executionController) ensureWorkflowExecutionRecord(ctx context.Context, exec *types.Execution, target *parsedTarget, payload []byte) {
-	workflowExec := c.buildWorkflowExecutionRecord(ctx, exec, target, payload)
-	if workflowExec == nil {
-		return
-	}
-
-	if err := c.store.StoreWorkflowExecution(ctx, workflowExec); err != nil {
-		logger.Logger.Error().
-			Err(err).
-			Str("execution_id", exec.ExecutionID).
-			Msg("failed to persist workflow execution state")
-	}
-}
-
-func (c *executionController) buildWorkflowExecutionRecord(ctx context.Context, exec *types.Execution, target *parsedTarget, payload []byte) *types.WorkflowExecution {
-	if exec == nil || target == nil {
-		return nil
-	}
-
-	runID := exec.RunID
-	if runID == "" {
-		runID = utils.GenerateRunID()
-	}
-
-	rootWorkflowID, parentWorkflowID, depth := c.deriveWorkflowHierarchy(ctx, exec)
-
-	startTime := exec.StartedAt
-	if startTime.IsZero() {
-		startTime = time.Now().UTC()
-	}
-
-	workflowName := fmt.Sprintf("%s.%s", exec.NodeID, exec.ReasonerID)
-	runIDCopy := runID
-	workflowExec := &types.WorkflowExecution{
-		WorkflowID:          runID,
-		ExecutionID:         exec.ExecutionID,
-		AgentFieldRequestID: utils.GenerateAgentFieldRequestID(),
-		RunID:               &runIDCopy,
-		SessionID:           exec.SessionID,
-		ActorID:             exec.ActorID,
-		AgentNodeID:         exec.AgentNodeID,
-		ParentWorkflowID:    parentWorkflowID,
-		ParentExecutionID:   exec.ParentExecutionID,
-		RootWorkflowID:      rootWorkflowID,
-		WorkflowDepth:       depth,
-		ReasonerID:          exec.ReasonerID,
-		Status:              string(exec.Status),
-		WorkflowName:        &workflowName,
-		StartedAt:           startTime,
-		CreatedAt:           startTime,
-		UpdatedAt:           startTime,
-		Notes:               []types.ExecutionNote{},
-	}
-
-	if len(payload) > 0 {
-		cloned := cloneBytes(payload)
-		workflowExec.InputData = json.RawMessage(cloned)
-		workflowExec.InputSize = len(cloned)
-	}
-
-	if target.TargetType != "" {
-		workflowExec.WorkflowTags = []string{target.TargetType}
-	} else {
-		workflowExec.WorkflowTags = []string{}
-	}
-
-	return workflowExec
-}
-
-func (c *executionController) deriveWorkflowHierarchy(ctx context.Context, exec *types.Execution) (*string, *string, int) {
-	runID := exec.RunID
-	rootWorkflowID := pointerString(runID)
-	var parentWorkflowID *string
-	depth := 0
-
-	if exec.ParentExecutionID != nil {
-		parentExecution, err := c.store.GetWorkflowExecution(ctx, *exec.ParentExecutionID)
-		if err != nil {
-			logger.Logger.Debug().
-				Err(err).
-				Str("execution_id", exec.ExecutionID).
-				Str("parent_execution_id", *exec.ParentExecutionID).
-				Msg("failed to load parent workflow execution")
-		}
-		if parentExecution != nil {
-			parentWorkflowID = pointerString(parentExecution.WorkflowID)
-			if parentExecution.RootWorkflowID != nil {
-				rootWorkflowID = parentExecution.RootWorkflowID
-			} else {
-				rootWorkflowID = pointerString(parentExecution.WorkflowID)
-			}
-			depth = parentExecution.WorkflowDepth + 1
-		} else {
-			depth = 1
-		}
-	}
-
-	return rootWorkflowID, parentWorkflowID, depth
-}
-
-func (c *executionController) updateWorkflowExecutionFinalState(
-	ctx context.Context,
-	executionID string,
-	status types.ExecutionStatus,
-	result []byte,
-	elapsed time.Duration,
-	errorMessage *string,
-) {
-	err := c.store.UpdateWorkflowExecution(ctx, executionID, func(current *types.WorkflowExecution) (*types.WorkflowExecution, error) {
-		if current == nil {
-			return nil, fmt.Errorf("execution with ID %s not found", executionID)
-		}
-		now := time.Now().UTC()
-		current.Status = string(status)
-		current.UpdatedAt = now
-		completedAt := now
-		current.CompletedAt = &completedAt
-		duration := elapsed.Milliseconds()
-		current.DurationMS = &duration
-		if len(result) > 0 {
-			cloned := cloneBytes(result)
-			current.OutputData = json.RawMessage(cloned)
-			current.OutputSize = len(cloned)
-		} else {
-			current.OutputData = nil
-			current.OutputSize = 0
-		}
-		current.ErrorMessage = errorMessage
-		return current, nil
-	})
-	if err != nil {
-		logger.Logger.Error().
-			Err(err).
-			Str("execution_id", executionID).
-			Msg("failed to update workflow execution state")
-	}
-}
-
-func cloneBytes(src []byte) []byte {
-	if src == nil {
-		return nil
-	}
-	dst := make([]byte, len(src))
-	copy(dst, src)
-	return dst
-}
-
-// callError wraps an upstream agent HTTP error, preserving the original status
-// code and response body for structured error propagation.
-type callError struct {
-	statusCode int
-	message    string
-	body       []byte
-}
-
-func (e *callError) Error() string {
-	return e.message
-}
-
-func writeExecutionError(ctx *gin.Context, err error) {
-	if err == nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "unknown error", "error_category": string(ErrorCategoryInternal)})
-		return
-	}
-
-	var ce *callError
-	if errors.As(err, &ce) {
-		category := classifyCallError(ce, err)
-		response := gin.H{
-			"error":          ce.message,
-			"error_category": string(category),
-			"status":         "failed",
-		}
-		// Preserve structured error data from the agent's response body.
-		if len(ce.body) > 0 {
-			var parsed interface{}
-			if json.Unmarshal(ce.body, &parsed) == nil {
-				response["error_details"] = parsed
-			}
-		}
-		// Propagate 4xx status codes from the agent (client-facing errors);
-		// use 502 Bad Gateway for 5xx (upstream server failure).
-		httpStatus := http.StatusBadGateway
-		if ce.statusCode >= 400 && ce.statusCode < 500 {
-			httpStatus = ce.statusCode
-		}
-		ctx.JSON(httpStatus, response)
-		return
-	}
-
-	var pe *executionPreconditionError
-	if errors.As(err, &pe) {
-		body := gin.H{
-			"error":          pe.Error(),
-			"error_category": string(pe.Category()),
-		}
-		// When a stable machine code is set, promote it to `error` and move
-		// the human-readable text to `message` — matching the contract used
-		// by reasoners.go / skills.go / permission middleware.
-		if code := pe.ErrorCode(); code != "" {
-			body["error"] = code
-			body["message"] = pe.Error()
-		}
-		ctx.JSON(pe.HTTPStatusCode(), body)
-		return
-	}
-
-	// Classify untyped errors (timeouts, connection failures, etc.)
-	category := classifyRawError(err)
-	httpStatus := http.StatusBadRequest
-	if category == ErrorCategoryAgentTimeout || category == ErrorCategoryAgentUnreachable {
-		httpStatus = http.StatusGatewayTimeout
-	}
-	ctx.JSON(httpStatus, gin.H{
-		"error":          err.Error(),
-		"error_category": string(category),
-	})
-}
-
-// classifyExecutionError determines the error category from any execution error.
-func classifyExecutionError(err error) ErrorCategory {
-	if err == nil {
-		return ErrorCategoryInternal
-	}
-
-	var ce *callError
-	if errors.As(err, &ce) {
-		return classifyCallError(ce, err)
-	}
-
-	var pe *executionPreconditionError
-	if errors.As(err, &pe) {
-		return pe.Category()
-	}
-
-	return classifyRawError(err)
-}
-
-// classifyCallError determines the error category for an agent call error.
-func classifyCallError(ce *callError, original error) ErrorCategory {
-	if ce.statusCode >= 500 {
-		return ErrorCategoryAgentError
-	}
-	if ce.statusCode == 408 {
-		return ErrorCategoryAgentTimeout
-	}
-	// Check if the body is valid JSON — if not, it's a bad response
-	if len(ce.body) > 0 {
-		var js json.RawMessage
-		if json.Unmarshal(ce.body, &js) != nil {
-			return ErrorCategoryBadResponse
-		}
-	}
-	return ErrorCategoryAgentError
-}
-
-// classifyRawError inspects an untyped error for timeout/connection patterns.
-func classifyRawError(err error) ErrorCategory {
-	if err == nil {
-		return ErrorCategoryInternal
-	}
-
-	errStr := err.Error()
-
-	// Context deadline exceeded = timeout
-	if errors.Is(err, context.DeadlineExceeded) || strings.Contains(errStr, "context deadline exceeded") {
-		return ErrorCategoryAgentTimeout
-	}
-
-	// Connection refused / reset = agent unreachable
-	if strings.Contains(errStr, "connection refused") ||
-		strings.Contains(errStr, "connection reset") ||
-		strings.Contains(errStr, "no such host") ||
-		strings.Contains(errStr, "i/o timeout") {
-		return ErrorCategoryAgentUnreachable
-	}
-
-	// Cancelled context
-	if errors.Is(err, context.Canceled) || strings.Contains(errStr, "context canceled") {
-		return ErrorCategoryInternal
-	}
-
-	// An agent or reasoner that does not exist. Matched on the messages
-	// prepareExecutionForTarget and determineTargetType produce, in the same
-	// style as the transport patterns above.
-	if strings.Contains(errStr, "not found") &&
-		(strings.HasPrefix(errStr, "agent '") || strings.HasPrefix(errStr, "target '")) {
-		return ErrorCategoryTargetNotFound
-	}
-
-	return ErrorCategoryInternal
-}
-
-// httpStatusForFailedExecution determines the HTTP status code to return to
-// callers for a failed execution record. It replaces the hardcoded 502 in the
-// async-completion branch with context-aware classification:
-//
-//  1. If StatusReason encodes a client error ("agent_client_error:<code>"), use that code.
-//  2. If StatusReason maps to a known server-side category, use the appropriate 5xx.
-//  3. Parse ErrorMessage for the "agent error (NNN):" pattern produced by the sync lane.
-//  4. Default to 502 Bad Gateway.
-func httpStatusForFailedExecution(exec *types.Execution) int {
-	// Check StatusReason for an encoded client error status code.
-	if exec.StatusReason != nil {
-		reason := *exec.StatusReason
-		if strings.HasPrefix(reason, "agent_client_error:") {
-			codeStr := strings.TrimPrefix(reason, "agent_client_error:")
-			if code, err := strconv.Atoi(codeStr); err == nil && code >= 400 && code < 500 {
-				return code
-			}
-		}
-		// Map known server-side categories.
-		switch ErrorCategory(reason) {
-		case ErrorCategoryAgentTimeout:
-			return http.StatusGatewayTimeout
-		case ErrorCategoryAgentUnreachable:
-			return http.StatusBadGateway
-		case ErrorCategoryTargetNotFound:
-			return http.StatusNotFound
-		case ErrorCategoryNodeUnavailable:
-			return http.StatusServiceUnavailable
-		case ErrorCategoryConcurrencyLimit:
-			return http.StatusTooManyRequests
-		}
-	}
-
-	// Fallback: parse ErrorMessage for the "agent error (NNN):" pattern that
-	// the sync lane produces when the agent returns an HTTP error directly.
-	if exec.ErrorMessage != nil {
-		msg := *exec.ErrorMessage
-		if strings.HasPrefix(msg, "agent error (") {
-			if idx := strings.Index(msg, "):"); idx > 13 {
-				codeStr := msg[13:idx]
-				if code, err := strconv.Atoi(codeStr); err == nil && code >= 400 && code < 500 {
-					return code
-				}
-			}
-		}
-	}
-
-	return http.StatusBadGateway
-}
-
-func pointerTime(t time.Time) *time.Time {
-	return &t
-}
-
-func pointerString(v string) *string {
-	return &v
-}
-
-func pointerInt64(v int64) *int64 {
-	return &v
-}
-
-func truncateForLog(body []byte) string {
-	const limit = 1024
-	if len(body) <= limit {
-		return string(body)
-	}
-	return string(body[:limit]) + "..."
-}
-
-const (
-	// bodyDigestPrefixLen is how many hex characters of the keyed body digest
-	// reach the log. Long enough that two distinct bodies colliding is not a
-	// practical concern for correlation.
-	bodyDigestPrefixLen = 16
-
-	// maxLoggedContentTypeLen bounds the agent-supplied content type. Response
-	// headers are attacker-influenced and Go accepts them up to megabytes; the
-	// log line must stay a log line.
-	maxLoggedContentTypeLen = 128
-)
-
-// bodyDigestKey is a random key minted once per process for the redacted body
-// digests below. A bare hash of the body would be a guessable commitment to it:
-// a short, low-entropy response — a one-time code, an email address, a bare
-// token, "true" — could be recovered offline by hashing candidates and matching
-// the logged digest, with body_bytes to prune the search. Keying the digest
-// removes that while keeping the property operators actually use, namely that
-// the same body logs the same digest within one run of the control plane.
-var bodyDigestKey = sync.OnceValue(func() []byte {
-	key := make([]byte, 32)
-	if _, err := rand.Read(key); err != nil {
-		// No entropy source: fail closed and log no digest at all rather than
-		// fall back to an unkeyed one.
-		return nil
-	}
-	return key
-})
-
-// redactedBodyDigest returns the logged digest prefix for a body, or "" when no
-// key could be minted.
-func redactedBodyDigest(body []byte) string {
-	key := bodyDigestKey()
-	if key == nil {
-		return ""
-	}
-	mac := hmac.New(sha256.New, key)
-	_, _ = mac.Write(body)
-	return hex.EncodeToString(mac.Sum(nil))[:bodyDigestPrefixLen]
-}
-
-// logSafeContentType reduces an agent-supplied Content-Type header to something
-// bounded: the media type without its parameters, truncated as a last resort.
-func logSafeContentType(contentType string) string {
-	if mediaType, _, err := mime.ParseMediaType(contentType); err == nil && mediaType != "" {
-		contentType = mediaType
-	}
-	if len(contentType) > maxLoggedContentTypeLen {
-		return contentType[:maxLoggedContentTypeLen] + "..."
-	}
-	return contentType
-}
-
-// annotateBodyForLog describes an agent response body on a log event.
-//
-// Agent responses are caller data, so they follow the same redaction switch as
-// execution payloads (logging.redact_payloads / AGENTFIELD_LOG_REDACT_PAYLOADS,
-// see SetRedactPayloads); redact is the caller's view of that switch. With
-// redaction on — the default — only non-reversible metadata is attached: the
-// response media type, the body length in bytes and a keyed digest prefix. That
-// is enough to recognise "same body as before" and to find the full payload in
-// the database, without the body itself reaching stdout. With redaction
-// explicitly disabled the previous truncated preview is attached instead.
-func annotateBodyForLog(event *zerolog.Event, contentType string, body []byte, redact bool) *zerolog.Event {
-	if event == nil {
-		return nil
-	}
-	if mediaType := logSafeContentType(contentType); mediaType != "" {
-		event = event.Str("content_type", mediaType)
-	}
-	event = event.Int("body_bytes", len(body))
-	if !redact {
-		return event.Str("body", truncateForLog(body))
-	}
-	event = event.Bool("body_redacted", true)
-	if digest := redactedBodyDigest(body); digest != "" {
-		event = event.Str("body_digest", digest)
-	}
-	return event
-}
-
-func (c *executionController) savePayload(ctx context.Context, data []byte) *string {
-	if c.payloads == nil || len(data) == 0 {
-		return nil
-	}
-	record, err := c.payloads.SaveBytes(ctx, data)
-	if err != nil {
-		logger.Logger.Warn().Err(err).Int("bytes", len(data)).Msg("failed to persist payload; proceeding without URI")
-		return nil
-	}
-	uri := record.URI
-	return &uri
-}
-
-func (j asyncExecutionJob) process() {
-	// Release the per-agent concurrency slot when this job finishes
-	if j.plan.target != nil {
-		defer ReleaseExecutionSlot(j.plan.target.NodeID)
-	}
-
-	// Use a bounded context so that paused executions do not block goroutines
-	// indefinitely if the resume/cancel event is never delivered (e.g. event bus
-	// crash, server restart). 24 hours is generous but prevents permanent leaks.
-	bgCtx, cancel := context.WithTimeout(context.Background(), 24*time.Hour)
-	defer cancel()
-
-	currentExec, err := j.controller.store.GetExecutionRecord(bgCtx, j.plan.exec.ExecutionID)
-	if err == nil && currentExec != nil {
-		if currentExec.Status == types.ExecutionStatusCancelled {
-			logger.Logger.Info().
-				Str("execution_id", j.plan.exec.ExecutionID).
-				Msg("skipping async agent call; execution cancelled")
-			return
-		}
-		if currentExec.Status == types.ExecutionStatusPaused {
-			if waitErr := j.controller.waitForResume(bgCtx, j.plan.exec.ExecutionID); waitErr != nil {
-				logger.Logger.Info().
-					Str("execution_id", j.plan.exec.ExecutionID).
-					Err(waitErr).
-					Msg("aborting async agent call while paused")
-				return
-			}
-		}
-	}
-
-	resultBody, elapsed, asyncAccepted, callErr := j.controller.callAgent(bgCtx, &j.plan)
-	if callErr == nil && asyncAccepted {
-		logger.Logger.Info().
-			Str("execution_id", j.plan.exec.ExecutionID).
-			Msg("agent accepted execution for async processing")
-		return
-	}
-
-	// Extract, persist (best-effort), and strip token/cost usage from the
-	// synchronous result envelope so it never leaks into the stored payload.
-	if callErr == nil {
-		if usageRaw, stripped := extractUsageFromResult(resultBody); usageRaw != nil {
-			resultBody = stripped
-			j.controller.ingestUsage(bgCtx, j.plan.exec, usageRaw)
-		}
-	}
-
-	job := completionJob{
-		controller: j.controller,
-		plan:       &j.plan,
-		result:     resultBody,
-		elapsed:    elapsed,
-		callErr:    callErr,
-	}
-	if err := enqueueCompletion(job); err != nil {
-		logger.Logger.Error().
-			Err(err).
-			Str("execution_id", j.plan.exec.ExecutionID).
-			Msg("failed to enqueue completion job for async execution")
-		if callErr != nil {
-			if updateErr := j.controller.failExecution(bgCtx, &j.plan, callErr, elapsed, resultBody); updateErr != nil {
-				logger.Logger.Error().
-					Err(updateErr).
-					Str("execution_id", j.plan.exec.ExecutionID).
-					Msg("fallback async failure persistence failed")
-			}
-		} else {
-			if updateErr := j.controller.completeExecution(bgCtx, &j.plan, resultBody, elapsed); updateErr != nil {
-				logger.Logger.Error().
-					Err(updateErr).
-					Str("execution_id", j.plan.exec.ExecutionID).
-					Msg("fallback async completion persistence failed")
-			}
-		}
-	}
-}
-
-func newAsyncWorkerPool(workerCount, queueCapacity int) *asyncWorkerPool {
-	pool := &asyncWorkerPool{
-		queue: make(chan asyncExecutionJob, queueCapacity),
-	}
-
-	for i := 0; i < workerCount; i++ {
-		go func(workerID int) {
-			for job := range pool.queue {
-				job.process()
-			}
-		}(i)
-	}
-
-	logger.Logger.Info().
-		Int("workers", workerCount).
-		Int("queue_capacity", queueCapacity).
-		Msg("async execution worker pool initialized")
-
-	return pool
-}
-
-func (p *asyncWorkerPool) submit(job asyncExecutionJob) bool {
-	select {
-	case p.queue <- job:
-		return true
-	default:
-		return false
-	}
-}
-
-func getAsyncWorkerPool() *asyncWorkerPool {
-	asyncPoolOnce.Do(func() {
-		workerCount := resolveIntFromEnv("AGENTFIELD_EXEC_ASYNC_WORKERS", runtime.NumCPU())
-		if workerCount <= 0 {
-			workerCount = runtime.NumCPU()
-		}
-
-		queueCapacity := resolveIntFromEnv("AGENTFIELD_EXEC_ASYNC_QUEUE_CAPACITY", 1024)
-		if queueCapacity <= 0 {
-			queueCapacity = 1024
-		}
-
-		asyncPool = newAsyncWorkerPool(workerCount, queueCapacity)
-	})
-	return asyncPool
-}
-
-func resolveIntFromEnv(key string, fallback int) int {
-	raw := strings.TrimSpace(os.Getenv(key))
-	if raw == "" {
-		return fallback
-	}
-	value, err := strconv.Atoi(raw)
-	if err != nil {
-		logger.Logger.Warn().
-			Str("key", key).
-			Str("value", raw).
-			Msg("invalid integer environment override; using fallback")
-		return fallback
-	}
-	return value
-}
-
-func ensureCompletionWorker() {
-	completionOnce.Do(func() {
-		size := resolveIntFromEnv("AGENTFIELD_EXEC_COMPLETION_QUEUE", 2048)
-		if size <= 0 {
-			size = 2048
-		}
-		completionQueue = make(chan completionJob, size)
-		go func() {
-			for job := range completionQueue {
-				err := processCompletionJob(job)
-				if job.done != nil {
-					job.done <- err
-					close(job.done)
-				}
-			}
-		}()
-	})
-}
-
-func processCompletionJob(job completionJob) error {
-	ctx := context.Background()
-	if job.callErr != nil {
-		return job.controller.failExecution(ctx, job.plan, job.callErr, job.elapsed, job.result)
-	}
-	return job.controller.completeExecution(ctx, job.plan, job.result, job.elapsed)
-}
-
-func enqueueCompletion(job completionJob) error {
-	ensureCompletionWorker()
-	select {
-	case completionQueue <- job:
-		return nil
-	default:
-		return fmt.Errorf("completion queue is full")
-	}
-}
